fix: arm Kl 64 位除法/取余与 isel store 常量处理
mcc arm 后端:
- arm_isel.c sel(): store 指令的 argcls 用 Ke 通配类，会让 fixarg 把
  整数 store 常量误当浮点 stash（storel 1000000000 变 R16=s load
  .Lfp0; storeb）。参照 aarch64，对 isstore 显式传 i.cls/Kl。
  同时捕获 curi->arg 指针（fixarg 会 emit 新指令移动 curi）。

meuos-libc (arm):
- stdint.h: 非 x86 目标 int64_t/uint64_t/int_fast64_t 改用 long long。
  arm 是 ILP32，long 为 32 位，原定义导致 aeabi.o 的 64 位参数被
  编译器按 32 位处理。
- aeabi.c udivmod64: 修正两处长除算法 bug —— R 左移时 rh 取 rl 顶位
  (原误取 nl 顶位)；nl 左移去掉错误的 &0x7fffffff 掩码（该位应流入
  nh）。修复后 100/7=14、3000000000/1000000000=3。
- aeabi_wrap.S: __aeabi_[u]ldivmod 调用 C helper 时未把 rem 指针存入
  栈参数槽（第 3 个参数在 r0-r3 之后于 [sp]），导致余数写入垃圾地址。
  补 str r4,[sp]。

+++ projects/meuos-libc/include/stdint.h
@@ -15,8 +15,10 @@ typedef unsigned long long uint64_t;
 typedef int intptr_t;
 typedef unsigned int uintptr_t;
 #else
-typedef long int64_t;
-typedef unsigned long uint64_t;
+/* ARM (ILP32) has 32-bit long; on the 64-bit targets long is 64-bit,
+ * but long long is 64-bit everywhere, so use it for the 64-bit types. */
+typedef long long int64_t;
+typedef unsigned long long uint64_t;
 typedef long intptr_t;
 typedef unsigned long uintptr_t;
 #endif
@@ -34,8 +36,8 @@ typedef long int_fast16_t;
 typedef unsigned long uint_fast16_t;
 typedef long int_fast32_t;
 typedef unsigned long uint_fast32_t;
-typedef long int_fast64_t;
-typedef unsigned long uint_fast64_t;
+typedef long long int_fast64_t;
+typedef unsigned long long uint_fast64_t;
 
 #define SIZE_MAX ((size_t)-1)
